Fixed up study delete validation; moved file tests
